Move files to correct folders
